MCP notifications, tool error handling, x402 design docs
JSON-RPC: JsonRpcRequest (required id) vs JsonRpcNotification, chosen by
parse_message on the presence of "id". Notifications never get a reply
(stdio: nothing, HTTP: 202); IGNORED_NOTIFICATIONS lists the MCP client
notifications we accept and why ignoring each is within the spec.

Tool errors: domain failures (TeeSwapError) are attested isError results
over MCP and 404/422 over REST; anything else is -32603 / 500, logged, no
details. The stdio loop survives tool and encode failures. QuoteError
replaces bare ValueErrors (incl. all-zero outputs); a chain without an RPC
is NoRouteError; invoice errors carry readable messages.

Fixes: tools/list failed to encode (float schema bounds, now int); stdio
session regression test covers the handshake, an isError result, an
unexpected error and the session continuing.

Docs: X402.md documents the x402 MCP and HTTP transports as specified, the
quote/accept/execute tools with _x402 variants (funding chosen at quote
time), and Tool / PaidTool: paid tools return PaymentRequired, transports
translate; settlement left open. Facilitator docs corrected (/supported
only, no /health; int poll_interval). ARCHITECTURE no longer calls x402
middleware.

+++ teeswap/wire.py
@@ -59,8 +59,9 @@ class WireSchema:
     description: str
     format: str | None = None
     pattern: str | None = None
-    minimum: Decimal | None = None
-    maximum: Decimal | None = None
+    # ints, not floats: these end up in MCP inputSchema, which encode() must accept
+    minimum: int | None = None
+    maximum: int | None = None
 
 
 class Validated(HasToWire):
